task_1: refractoring

+++ src/by/epamtc/lonelyday/tasks_01/task1/service/DigitService.java
@@ -1,7 +1,7 @@
 package by.epamtc.lonelyday.tasks_01.task1.service;
 
 public class DigitService {
-    public static int calcLastIntDigit(int n){
-        return n % 10;
+    public static int calcLastIntDigit(long n){
+        return (int) (n % 10);
     }
 }
