+++ src/by/epamtc/lonelyday/tasks_01/task1/service/SquareService.java
@@ -1,12 +1,12 @@
 package by.epamtc.lonelyday.tasks_01.task1.service;
 
 public class SquareService {
-    public static int calcIntSquare(int n){
-        return n*n;
+    public static long calcIntSquare(int n){
+        return n*(long)n;
     }
 
     public static int calcIntSquareLastDigit(int n){
-        int square = calcIntSquare(n);
+        long square = calcIntSquare(n);
         return DigitService.calcLastIntDigit(square);
     }
 }
